requirements.txt
dash
plotly
pandas
numpy
json
dash-bootstrap-components-1.2.1
